Added UI per scene, follows previously added active states
