+++ Eden/Assets/Prefabs/Scenes/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff126e446c351bf458bf4ac1a373d05a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
