+++ Eden/Assets/Prefabs/Scenes/UI/scene1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c7433ef65e7c41e46a78d59d95a47b8c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
