+++ Eden/Assets/Prefabs/Scenes/UI/scene2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e7f7fa4f097d8d24d99c2c758d06f40f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
