+++ Eden/Assets/Prefabs/Scenes/UI/scene3.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6282faeb6060ba843894ed3c90f9a225
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
