+++ Eden/Assets/Scenes/LivingRoom.asset
@@ -14,7 +14,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   sceneName: LivingRoomEnd
   sceneIndex: 4
-  sceneUIPrefab: {fileID: 0}
+  sceneUIPrefab: {fileID: 3857556652320349115, guid: 57e21f0c8fd771e408614b342f6fb27f,
+    type: 3}
   sceneEnvironmentPrefab: {fileID: 8491744840830792481, guid: 59512f7ffe70b884ba15481e36c93fb0,
     type: 3}
   sceneState:
