+++ Eden/Assets/Scenes/LivingRoomScene.asset
@@ -14,7 +14,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   sceneName: LivingRoomBegin
   sceneIndex: 0
-  sceneUIPrefab: {fileID: 0}
+  sceneUIPrefab: {fileID: 4210348463103143527, guid: c7433ef65e7c41e46a78d59d95a47b8c,
+    type: 3}
   sceneEnvironmentPrefab: {fileID: 2480974826615126096, guid: f2586b83c3e5d6e489ca4ebf69816495,
     type: 3}
   sceneState:
