+++ Eden/Assets/Scenes/Market.asset
@@ -14,7 +14,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   sceneName: Market
   sceneIndex: 3
-  sceneUIPrefab: {fileID: 0}
+  sceneUIPrefab: {fileID: 4580075323178319539, guid: 08664117af567c5448448bde406cb50f,
+    type: 3}
   sceneEnvironmentPrefab: {fileID: 281836812921029424, guid: 1e2c1359fdb351f4887a8d59757d564f,
     type: 3}
   sceneState:
