--- Eden/Assets/Scripts/Interactable.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[Serializable]
-public class Interactable : MonoBehaviour
-{
-
-}
