+++ Eden/Assets/Scripts/Interactables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8d9baabf9376a8f40bd8143fcd7acf09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
