+++ Eden/Assets/Scripts/Interactables/Interactable.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[Serializable]
+public class Interactable : MonoBehaviour
+{
+
+}
