+++ Eden/Assets/Scripts/Interactables/Paper.cs.meta
