+++ Eden/Assets/Scripts/Interactables/Trash.cs.meta
