+++ Eden/Assets/Scripts/Managers/UIManager.cs
@@ -10,7 +10,18 @@ using UnityEngine.UI;
 /// </summary>
 public class UIManager : MonoBehaviour
 {
+    public List<GameObject> SceneUIs
+    {
+        get { return sceneUIs; }
+        set { sceneUIs = value; }
+    }
+
+    [SerializeField] private Transform uiParent;
+
     private static UIManager instance;
+
+    private List<GameObject> sceneUIs;
+    private Dictionary<string, GameObject> spawnedUIs;
 
     public static UIManager Instance
     {
@@ -31,7 +42,7 @@ public class UIManager : MonoBehaviour
         }
     }
 
-    private void Awake()
+    protected void Awake()
     {
         if (instance != null && instance != this)
         {
@@ -43,4 +54,39 @@ public class UIManager : MonoBehaviour
             DontDestroyOnLoad(this.gameObject);
         }
     }
+
+    /// <summary>
+    /// Method to spawn all spawnable prefabs, set their name and add them to
+    /// a list to keep track.
+    /// </summary>
+    public void SetSpawnablePrefabs()
+    {
+        spawnedUIs = new Dictionary<string, GameObject>();
+
+        foreach (GameObject prefab in sceneUIs)
+        {
+            GameObject newPrefab = Instantiate(prefab, uiParent);
+            newPrefab.name = prefab.name;
+            spawnedUIs.Add(prefab.name, newPrefab);
+
+            newPrefab.SetActive(false);
+        }
+    }
+
+    public void ShowUI(List<GameScene> scenes, string sceneName)
+    {
+        foreach (GameScene scene in scenes)
+        {
+            foreach(var entry in spawnedUIs)
+            {
+                entry.Value.SetActive(false);
+
+                if (entry.Key == sceneName)
+                {
+                    entry.Value.SetActive(true);
+                }
+            }
+        }
+    }
+
 }
