--- Eden/Assets/Scripts/Trash.cs
@@ -1,28 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-
-public class Trash : Interactable
-{
-    public int Score => score;
-    
-    public TrashType trashType;
-
-    [SerializeField] private int score = 1;
-
-    public enum TrashType
-    {
-        Bottle,
-        Can,
-        Bag,
-    }
-
-    public bool PickUp()
-    {
-        Destroy(gameObject);
-
-        return true;
-    }
-}
